// src/app/shared/models/userRegistrationDto.ts
import { Guid } from 'guid-typescript';

export class UserRegistrationDto {
    Role: string;
    FirstName: string;
    LastName: string;
    Email: string;
    PhoneNumber: string;
    Pesel: string;
    PostCode: string;
    City: string;
    Street: string;
    HouseNumber: string;
    Password: string;
    MilitaryRank:string;
}